move and consolidate EntityList and types in base folder
