+++ assets/src/application/ui/layout/entityList/index.ts
@@ -1,4 +1,4 @@
 export { default as EntityList } from './EntityList';
-export * from './EntityList/types';
 export * from './filterBar';
 export * from './pagination';
+export * from './types';
